*** empty log message ***

+++ trunk/gsdx9/GSTables.h
@@ -21,16 +21,16 @@
 
 #pragma once
 
-extern WORD blockTable32[4][8];
-extern WORD blockTable32Z[4][8];
-extern WORD blockTable16[8][4];
-extern WORD blockTable16S[8][4];
-extern WORD blockTable16Z[8][4];
-extern WORD blockTable16SZ[8][4];
-extern WORD blockTable8[4][8];
-extern WORD blockTable4[8][4];
-extern WORD columnTable32[8][8];
-extern WORD columnTable16[8][16];
-extern WORD columnTable8[16][16];
-extern WORD columnTable4[16][32];
+extern DWORD blockTable32[4][8];
+extern DWORD blockTable32Z[4][8];
+extern DWORD blockTable16[8][4];
+extern DWORD blockTable16S[8][4];
+extern DWORD blockTable16Z[8][4];
+extern DWORD blockTable16SZ[8][4];
+extern DWORD blockTable8[4][8];
+extern DWORD blockTable4[8][4];
+extern DWORD columnTable32[8][8];
+extern DWORD columnTable16[8][16];
+extern DWORD columnTable8[16][16];
+extern DWORD columnTable4[16][32];
